generic/tests/os_update: refactor and have MultiVM support
yum_update.py had support to perform os update for RHEL based guests,
refactor it to support Ubuntu and adopt MultiVM. The patch also renames
yum_update.py -> os_update.py similarly cfg file is renamed.
