+++ generic/tests/cfg/os_update.cfg
@@ -1,0 +1,7 @@
+- os_update: install setup image_copy unattended_install.cdrom
+    virt_test_type = qemu libvirt
+    no Windows
+    no SLES
+    type = os_update
+    os_update_timeout = 600
+    shell_prompt = r"[\#\$]\s*$"
