--- generic/tests/cfg/yum_update.cfg
@@ -1,6 +1,0 @@
-- yum_update: install setup image_copy unattended_install.cdrom
-    virt_test_type = qemu libvirt
-    only Fedora, RHEL, CentOS
-    type = yum_update
-    shell_prompt = "Is this ok"
-    yum_update_timeout = 600
